+++ BaseListener.Tests/LogCallAspectFixtureCollection.cs
@@ -1,0 +1,9 @@
+using Hackney.Core.Testing.Shared;
+using Xunit;
+
+namespace BaseListener.Tests
+{
+    [CollectionDefinition("LogCall collection")]
+    public class LogCallAspectFixtureCollection : ICollectionFixture<LogCallAspectFixture>
+    { }
+}
